# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "APIService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/catch-me": {
      "post": {
        "operationId": "APIService_CatchMeIfYouCan",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCatchMeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCatchMeRequest"
            }
          }
        ],
        "tags": [
          "APIService"
        ]
      }
    },
    "/v1/summary": {
      "get": {
        "operationId": "APIService_GetSummary",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiSummaryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "APIService"
        ]
      }
    },
    "/v1/triangle/loop": {
      "get": {
        "operationId": "APIService_GetTriangleWithLoop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTriangleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "APIService"
        ]
      }
    },
    "/v1/triangle/non-loop": {
      "get": {
        "operationId": "APIService_GetTriangleWithNonLoop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTriangleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "APIService"
        ]
      }
    }
  },
  "definitions": {
    "apiCatchMeRequest": {
      "type": "object",
      "properties": {
        "input": {
          "type": "string",
          "title": "Single string input"
        }
      }
    },
    "apiCatchMeResponse": {
      "type": "object",
      "properties": {
        "output": {
          "type": "string"
        }
      }
    },
    "apiSummaryResponse": {
      "type": "object",
      "properties": {
        "beef": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "apiTriangleResponse": {
      "type": "object",
      "properties": {
        "sum": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}